FEM: results, move vonMises from result reader to result tools and use new result stress float lists
